Change: rename files as per sonarqube (#576)
